Update to Android 11.0.0 Release 37
* May Security update
* Still android11-qpr2 (didn't switch our branches in March because
  changes merged fine)

Build IDs:
* RQ2A.210505.003 for the Pixels 5, 4a (5G) - android-11.0.0_r37
* RQ2A.210505.002 for the Pixels 3, 3a, 4, 4a - android-11.0.0_r36

There are no other changes between the tags so we're just using r37 for all

Kernels and kernel modules:
* android-11.0.0_r0.78 for the Pixels 3, 3a
* android-11.0.0_r0.79 for the Pixel 4
* android-11.0.0_r0.80 for the Pixel 4a
* android-11.0.0_r0.81 for the Pixels 5, 4a (5G)

Note: Back to fts_touch_s5 for the 4a (from fts_touch)
Why they keep going back and forth, I have no idea!
The worst part is that fts_touch still has a android-11.0.0_r0.80 tag
but that's wrong and actually the Pixel 4 tag - so it syncs and builds
but touchscreen won't work with that. Sad, found this out the hard way :(

Change-Id: I181ee28111a4d59839fa983cd592ee50f523b0b2