manifest: Track custom prebuilts/misc revision
AOSP added only the arm version of libprotobuf-cpp-{lite,full}-21.12
but some of our older vendor blobs also depend on the arm64 version.
We can't fork prebuilts/misc due to its size, so do this trick to add
the arm64 prebuilts of libprotobuf-cpp-{lite,full}-21.12 since we have
no other way of overriding the build target or the repository.

Change-Id: Idde1e5fb3cbc3c260fe629a84afced1e17231287